Add type signatures to describeSymbol()
Summary:
Extremely useful to know the type of the symbol when calling describe. Currently the codehub api page scrapes the symbols' locations' documentSymbols() to find the signature attribute. We can just put it here.

The C++ and Hack type signature code assumes the entity has been recursively expanded.
That's not true in searchEntity/runSearch, was doing a shallow search only.

We can do a few things:
- since we only want the first result, or log it, do depth =2
- recursively expand

Rest is easy.

Reviewed By: bochko

Differential Revision: D38778030

fbshipit-source-id: 17c11dea36d2c7448a28a021c1de78bbb22fa772