Default to reading the schema index
Summary:
All prod servers are already using this, so it only affects starting
the server (or other Glean tools with `--db-root`) manually from
command line.

Only downside is the shell takes longer to start with `--db-root` but
normally you want to use it with `--schema` anyway which reads the
schema from the local filesystem.

Reviewed By: donsbot

Differential Revision: D38619322

fbshipit-source-id: 97a54ca60262259f728ab00528c2bd5e6a579330