Fix handilng of built-in functions.
Summary:
The first use of a built-in function seems to be considered a declaration. This leads to varying versions of `cxx1.Declarations` for a file depending on the include order and whether this is the first time built-in function is seen.

This diff changes the handling of built-in functions such as `__builtin_is_constant_evaluated` to not be declared, as well as remove the xrefs to them.

Reviewed By: malanka

Differential Revision: D38823647

fbshipit-source-id: 874f8ef34d1d06e527ea28a9689825206988189a