Make glass SymbolDescriptions a non-empty list of locations
Summary:
Instead of assuming we can describe a symbol id with a single location,
we expand this to take optionally more locations (i.e. a NonEmpty).

For cases like Hack namespaces ,where we will correctly match multiple source
locations in the same repo, the sym_locations field will have the extra locations.

This just adds the type and starts doing the plumbing. The extra locations are
not set yet. We already have a list of comments and annotations, so no need to extend those.

- we will retire `locations`, which uses the unnecessary SymbolPath type (just use LocationRange)

Reviewed By: nhawkes

Differential Revision: D38855344

fbshipit-source-id: 248ed06ed38201acc66b4f14b9ffd3db1cf98b77