Link simple arguments to Angle query.
Summary: Until now, the argument of the function annotated with the `angle_query` decorator was considered the query argument by default. Queries can have multiple arguments; and sometimes it is easier to use a literal. For these reasons, I had to make sure that there is a connection between the the argument of the decorated function and the argument of the queried entity. This diff is a preparatory step for adding record type support.

Reviewed By: nhawkes

Differential Revision: D38871971

fbshipit-source-id: 0839eaf183db50ca23e3917e942951a4508d9c44