Remove flaky test
Summary:
This test does describe on a namespace symbol that's split over two files.
Depending on the indexer though it picks one or the other as the anchor location. Making it non-det
We fix this by removing the old `location` field and then some ordering for sym_location and sym_other_locations

But for now don't block conveyor

Differential Revision: D39045771

fbshipit-source-id: bebd131183e847712fb8d251056c6a710c636a2a