Optimise existence checks
Summary:
If the variables bound by a group of statements are not needed to build the query
results, then iterating more than once through the results of these statement is
wasteful given all repeated executions will yield the same result.
These statements are there just to check that some condition is fulfilled.

Consider the query
```
P { a: { b: { c: X } } };
Q X;
R _;
```
It means that if the content of the database satisfies the condition specified by
the statements 1 and 2, then collect all values of `R`.
Now, it doesn't matter how many values of `P` and `Q` satisfy the condition and
it would be redundant to execute `Q _` for each match of `P` and `Q`.
Therefore, we could make this more efficient by saying that whatever comes after
lines 1 and 2, which make this existence check, should only be run once.

We could add double negations around theses statements in the source
to ensure we stop after the first result.  But double negation is rather cryptic
and not a nice thing to have in our schemas.  So instead of adding that to the
source, Glean could identify those statements and optimise them accordingly.

We can implement this with an optimisation pass after `fixBindings` which would
identify theses statements by traversing a reverse dependency graph of statements
from the query result. Groups of statements not in this path can be optimised.

The check implemented in this diff is much simpler though. It simply states that
if a statement doesn't bind any variable at all, then all repeated executions of
subsequent statements will be identical and therefore the statement can be
optimised to return a single result.
It misses some statements that would be suitable candidates and doesn't
handle groups of statements.
Statements such as `_ = P { X, X }`  where `X`  is not used anywhere afterwards
don't qualify in this implementation.

This will already improve some cases and I can implement the more comprehensive
check in the future.

Reviewed By: simonmar

Differential Revision: D39176391

fbshipit-source-id: 2984949c9ed2b62abba47270840f641387712971