Fix a bug in the Haskell codegen
Summary:
The RecordFields and SumFields type instances were wrong for predicate
types nested inside `maybe` or array types. In those cases we used the
key type instead of the predicate type.

In many cases it didn't matter, because these types are mostly used to
typecheck queries, and Angle allows key types and predicate types
interchangeably. However, if a query returns a non-predicate type,
then we also use that type to deserialize the query results, so the types
must not lie. It was possible to use the wrong deserializer with the
previous types.

The only thing you'll notice as a user is you may need to insert an
extra `predicate` inside a `just` (see the diffs to Glass for examples).

Reviewed By: donsbot

Differential Revision: D39171133

fbshipit-source-id: 642922c8a1da15b71166c7a5b30e81f314b70223