Return typechecked key size in register (#259)
Summary:
We used to pass in a pointer but returning this via a register seems better and will allow us to remove the `LoadWord` and `StoreWord` insns.

WARNING: This does *not* bump the bytecode version because I believe that `LoadWord` and `StoreWord` were never used in code generated for queries and hence never serialised.

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/259

Reviewed By: pepeiborra

Differential Revision: D39236399

Pulled By: donsbot

fbshipit-source-id: 1783d7b9dbbf8e86b8c180cae73c70f7240ab0f7