Serialise fact ids for prefix iterators
When we serialise iterators in query continuations, we're currently
including the key of the fact they're pointing to. This can be very
large. We want to serialise the fact id instead and then use that to
obtain the key when deserialising (via an additional `factById` query).
This commit does that in a backward and forward compatible way.

We serialise the fact id but also still serialise the key - this means
that old servers can deal with our continuations. When deserialising, we
use the fact id if present - thus, the code will still work when we stop
serialising keys altogether.