Serialise fact ids for prefix iterators (#263)
Summary:
When we serialise iterators in query continuations, we're currently
including the key of the fact they're pointing to. This can be very
large. We want to serialise the fact id instead and then use that to
obtain the key when deserialising (via an additional `factById` query).
This commit does that in a backward and forward compatible way.

We serialise the fact id but also still serialise the key - this means
that old servers can deal with our continuations. When deserialising, we
use the fact id if present - thus, the code will still work when we stop
serialising keys altogether.

NOTE: https://github.com/facebookincubator/Glean/issues/264 removes the actual key serialisation. I suggest merging this, updating all servers, and then merging https://github.com/facebookincubator/Glean/issues/264.

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/263

Reviewed By: simonmar

Differential Revision: D39246776

Pulled By: donsbot

fbshipit-source-id: a1ad135e54b8fe1410f1b62d492e37021e3b1dc3