Fixed xrefs should also resolve to their defn if possible
Summary: Fixed xrefs, we were always returning a decl. We should be checking if its a possible defn first. While this doesn't change the xref target, it does change the symbol id we will generate in the future to a defn occurrence, which is the canonical form. And this will make it simpler to do proper find-refs via the decl family of the defn.

Reviewed By: pepeiborra

Differential Revision: D39251774

fbshipit-source-id: 58024c8d0a1bf4212057bb1b2d0b2e2db7ba4f76