Expand C++ find-refs via decl families
Summary:
This generalizes the find-refs used by Glass for C++ to more uses, which appears to be "what clangd does" (when clangd has indexed the relevant files).

For decls or defns we lookup the decl family and return all uses thereof.

For defns, we have not been doing find-refs at all, which explains a few things.

- find base decl
- look up decl family
- do find-refs on them

## what's next ?

Arguably, we might also want to find the decl occurrences. Currently, we load the uses of the defn's decl family. But the decl family itself might be considered additional uses. ClangD seems to do this. To do this we probably need to return a RangeSpan as we'll have a mix of spans and ranges.

```
$ glass --tier glean.glass.dons find-references fbsource/cpp/fbcode/facebook/biggrep2/common/getShardId
fbsource/fbcode/codesearch/biggrep2/sync/BigGrepPackage.cpp 200:19-200:29
fbsource/fbcode/codesearch/biggrep2/BigGrep2MasterHandler.cpp 709:26-709:36
fbsource/fbcode/codesearch/biggrep2/BigGrep2MasterHandler.cpp 723:28-723:38
```

Reviewed By: simonmar

Differential Revision: D39252538

fbshipit-source-id: e649084bb9297eee69fe0fde4bdad72242c68231