Add USR hash for C++ decl into Glean
Summary:
Stores USR hash as hex string in Glean DB for function declarations.
See  https://fb.workplace.com/groups/1149405085538468/permalink/1403602193452088/ for the discussion and context.

Creates SHA1 hash of USR and uses 8 bytes of which as Index. This is same as what is being done in clangd. This would allow clangd to use Glean index as a global Index. We currently keep hex string (16bit encoding). We could use 64bit encoding to reduce the size further.

With the additional fact, Glean DB for the biggrep  has 1.8% overhead
```
test> :stat -s cxx1.USRDeclartion
cxx1.USRDeclartion.5
  count: 132225
  size:  8594023 (8.20 MB) 1.8678%
```

USR's for some of the functions in namespaces std, folly etc dumped as part of biggrep build can be found in P524573726

Reviewed By: simonmar

Differential Revision: D38938551

fbshipit-source-id: f3ff7d84b97894a2573cd7fe97b8dd1dd3c39442