Factor out the tree of include traces from PPTrace.
Summary:
This diff introduces a new predicate `cxx1.IncludeTree`, which factors out the traces that are part of `IncludeTrace`.
There are many variations of include traces that `cxx1.PPTrace` ends up with. Each of the variations cause duplications
of `[PPEvent]`. My suspicion is that most of the explosion however comes from the duplication of `pp1.Use`, based on
the current usage statistics.
```
pp1.Define.1
  count: 1890594
  size:  97803230 (93.27 MB) 0.0594%
pp1.Include.1
  count: 5799145
  size:  319459031 (304.66 MB) 0.1941%
pp1.Macro.1
  count: 782067
  size:  77275616 (73.70 MB) 0.0470%
pp1.Undef.1
  count: 65104
  size:  3348317 (3.19 MB) 0.0020%
pp1.Use.1
  count: 78934785
  size:  6253224497 (5.82 GB) 3.8001%
```
By factoring out the various of include traces out and reducing the duplication of references to the many many `pp1.Use`s,
we can significantly reduce the amount of total data.

`cxx1.IncludeTree` stores a `cxx1.Trace` and effectively a `[maybe cxx1.IncludeTree]` which corresponds to
the `pp1.Include` entries in `[cxx1.PPEvent]`. We could also add a `includes : [pp1.Include]` field to `cxx1.PPTrace`
if it would improve usability.

In this model, `cxx1.Trace` really represents a "fixed" parts of a file, and `cxx.IncludeTree` represents the "variable" includes of a file.

This change brings us from the initial starting point of 191 GB down to 153 GB, about a 20% decrease in DB size.

Reviewed By: simonmar

Differential Revision: D39156377

fbshipit-source-id: 11131adc526a59a09f6f3b529b38edee1bc5f568