Update API to support recursive creation of inner key types.
Summary:
Some union types contained records or unions as keys. The API is now supporting the correct generation of those types.
Some queries might fail at the moment, because the names for `EnumeratedTy` keys are generated but the types are never defined. This diff needs to land along with the API update for `EnumeratedTy`. I am making 2 diffs because logically they are distinct features.

 ---

Reviewed By: nhawkes

Differential Revision: D39305772

fbshipit-source-id: 4a157afbd2fe7563e3525322b92d5f139126ea21