Angle api for decoding cxx entities from symbol ids
Summary:
Many of the issues with C++ symbol ids come from trying to reuse the search api (search.cxx.SearchByScope/SearchByName) to do entity lookup.

Using search is the wrong approach because:
- it is lossy (we don't make everything searchable)
- it has wild cards and other approximations (so it tends to find more things)
- it has a loose relationship with decl and definitions

For code.cxx.Entity recovery from symbol ids we have a few things on our side:
- the full, complete scope (not wild cards needed)
- the fact it definitely exists, so we can assume lookup and not search
- tags to distinguish decls and defs
- we want _everything_ to be encodable, including exotic stuff

So this is step 1:
- expose all ways to extract a cxx1.Declaration by Scope-like things.
- use code.cxx.DeclToDef as the oracle for what is a "defn"

We should immediately be able to find more things, and more accurately.

Todo: we will use scope queries and similar things, but these are too generic I think, e.g. they permit "global" to appear in places it logically can't. We can likely improve the schema or the query to make the lookup cheaper.

Reviewed By: simonmar

Differential Revision: D39301673

fbshipit-source-id: a9a68e7ff679442f9abfd6d70e572deee0b08f64