Fix C++ attribute symbolParent self-reference, except for local varilables
Summary:
**First ever haskells for me so I might have done something silly.**

Fixes self-reference issues when making attribute symbolParent, except for local variables, as there is no straightforward way to do this with `Cxx.FunctionName`.

As such, stuff declared *IN* functions will not receive a parent attribute (which I think is still an improvement over it being declared with a parent itself).

In short the issue is for some declaration types (like type aliases) we displayed the symbol itself as its parent in the symbolParent property, like this (contrived):
```
"symbolSignature": {
   "aString": "using foo::bar::Symbol = Other;"
},
"symbolParent": {
  "aString": "Symbol"
},
```

Arguably we may need to display the full scope eventually, the diff brings the current state to be the following
```
"symbolSignature": {
   "aString": "using foo::bar::Symbol = Other;"
},
"symbolParent": {
  "aString": "bar"
},
```
But this is still incomplete info. It could be `foo::bar` or `std::bar` or `::bar` or whatever, and the user would never know unless they look at the signature.
Instead we may actually want
```
"symbolSignature": {
   "aString": "using foo::bar::Symbol = Other;"
},
"symbolParent": {
  "aString": "::foo::bar"
},
```
Up for discussion

Reviewed By: donsbot

Differential Revision: D39383241

fbshipit-source-id: 5da06a838c2c6d4d5a16876935cfe742c017f5ee