Fix bug in cxx1.DeclInRecord
Summary:
I think this has been broken for a long time. It also doesn't get used by anything significant (?) so likely unnoticed.

There's an invalid clause that causes nothing to match: the QName of the parent record is unified with the QName of the child. (QN == QN). That's wrong.

So this only ever worked for FunctionQNames, where FQN /= QN.  In fact, if you try to use it just times out, searching for records that contain themselves.

The fix is to rename the outer QN fact.

The impact is you couldn't use this to find the precise parent records of:

- variables
- records
- enums

only functions (and using declarations). Which might have been enough to pass the test but doesn't work when you see it on code

```
Scope{recordWithAccess={record=QN}} =
    (
      S where { usingDeclaration = X } = Decl;
      UsingDeclaration{name = FQN} = X;
      FunctionQName{scope = S} = FQN;
    ) | (
      S where { usingDirective = X } = Decl;
      UsingDirective{name = QN} = X;
      QName{scope = S} = QN
    ) | (
      S where { record_ = X } = Decl;
      RecordDeclaration{name = QN} = X;
      QName{scope = S} = QN
    ) | (
      S where { enum_ = X } = Decl;
      EnumDeclaration{name = QN} = X;
      QName{scope = S} = QN
    ) | (
      S where { function_ = X } = Decl;
      FunctionDeclaration{name = FQN} = X;
      FunctionQName{scope = S} = FQN;
    ) | (
      S where { variable = X } = Decl;
      VariableDeclaration{name = QN} = X;
      QName{scope = S} = QN
    ) | (
      S where { typeAlias = X } = Decl;
      TypeAliasDeclaration{name = QN} = X;
      QName{scope = S} = QN
    );
```

Reviewed By: nhawkes

Differential Revision: D39683445

fbshipit-source-id: c2a6f4921d5d43b5f53e5a0bc208c659f7bf1f80