Make matchPat efficiently bind the full key
Summary:
I want `withPrefix` to not need to return a function to capture the key.
But for that to be possible, `matchPat` should be able to efficiently
capture the full key without traversing it.

This is implemented here by observing that we can always copy the full
value without traversing it if we are in the last query chunk of a
branch that matches until the end of the input. Then we can just
copy from `input` to `inputend`.

This has the added benefit that we can now copy anything that is not a word
type at the end of the input without traversing its content.

Reviewed By: simonmar

Differential Revision: D39449973

fbshipit-source-id: 61909fdb71939144c03f7e52067d6ed48eb76d93