Add QueryPrefixWild
Summary:
Currently `withPrefix` returns a prefix plus the remaining chunks to be matched after the prefix portion.

This puts the responsibility of skipping the prefix before matching on the remaining chunks on the caller of `withPrefix`.

This setup will become cumbersome when we start transforming these prefixes. When a transformation occurs we will match the chunks from the beginning of the fact, regardless of the size of the prefix.

Using `QueryPrefixWild` we will be able to skip parts of the prefix from within `matchPat`, which means we can pass the suitable chunks to `matchPat` whether they match from the beginning or not.

Reviewed By: pepeiborra

Differential Revision: D39468065

fbshipit-source-id: 213c12b391c6a1133388dafb09d5a3dbf7d65b73