Support running multiple commands from CLI (#279)
Summary:
This adds a `glean script` command which reads commands (as they would be specified on the cmd line) from a file or stdin and executes them. The immediate use case for me is easier benchmarking (esp. in conjunction with `--db-memory` from https://github.com/facebookincubator/Glean/issues/274) but this seems generally quite useful to me and it's very easy to support.

See the example in `script --help` for how to use and how to test.

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/279

Reviewed By: lazamar

Differential Revision: D39761398

Pulled By: pepeiborra

fbshipit-source-id: 7ccacd6b25f5f336415eb99beaa41c55e95bea8b