Update Macro types with ByteSpans
Summary:
This diff does a couple of things. The main purpose is to make the range for function like macros match what you would see with actual functions. That is, the usage should only encompass the function name itself. To do that, a couple of things were needed.

## Adding additional fields to pp1.Use and pp1.Include

Both of these predicates were using the deprecated src.ByteRange. This is not compatible with glass. To glass, this looks the same as a ByteSpan. This adds in a ByteSpan that carries the same info for both structs. There are changes in preprocessor.cpp and db.cpp that fill in the new fields.

## Passing on the new fields cxx1.FilePPTraceXRefs

Glass calls `codemarkup.pp.PpFileEntityTraceXRefLocations` which ultimately retrieves the XRefs from cxx1.FilePPTraceXRefs. This passes on the new fields.

## Using the new fields in codemarkup.pp.PpFileEntityTraceXRefLocations

Here, we use the new field for macro defines, but not for includes or undefs. We don't want to change the behavior of includes or undefs, only for function like macro usages.

## EDIT for V4 and above

This diff now just makes the schema and indexer changes. The codemarkup changes will be coming on top of this.

Reviewed By: simonmar

Differential Revision: D39863161

fbshipit-source-id: 65d54b9b8009def6750024d6929b0fd2bfa0f455