Implement predicateStats for in-memory backend (#288)
Summary:
We want to make the in-memory backend work properly and this was a major missing piece. This adds support for computing `PredicateStats` to `FactSet` and threads it all the way through to the in-memory backend code. The stats are computed lazily and cached. There are a few knock-on effects, mostly due to moving `PredicateStats` from `rocksdb` to `rts`.

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/288

Reviewed By: simonmar

Differential Revision: D40014847

Pulled By: pepeiborra

fbshipit-source-id: 82689d5697f7c4d21607735a693d4e8873ba2f9d