Use (Typed Output) instead of Var in QueryChunks
Summary:
A `Var` references a binary output by its index in our vector of variables.
For schema transformations we might need to modify the contents of a variable
to be able to use it in the construction of a prefix (e.g. changing field
orders in a record). We want to keep the original variable as is, but use a
new binary output for the transformed version.
To avoid having to modify our vector of variables, let's keep the binary
outputs stored directly in the `QueryChunks`. This way we can easily
introduce new outputs using continuations.

Reviewed By: simonmar

Differential Revision: D39930806

fbshipit-source-id: b9929178b75cebcc07895dc1c05da4879c7e8788