Retrieve transformations applied from return type
Summary:
We only need to apply transformations relevant to the type being returned to the client, so if we know the type we have to return we can find out the transformations that will be required.

This removes the need to explicitly include the map of transformations applied in the continuation given it can be retrieved from the return type.

It also allows us to move the code to retrieve the transformations outside of `flatten` and into the body of `userQueryImpl`.

Reviewed By: simonmar

Differential Revision: D40060909

fbshipit-source-id: 2bca8cf4dac3aeea13072fa77ef85f1b333b86b6