Refactor FactSet's fact table (#290)
Summary:
This is one of a number of preparatory refactors before a big rewrite of `FactSet`'s implementation. We pull out the `Id`-based fact table that `FactSet` uses into a separate class whose interface is based on `Fact::Ref`, not `Fact*`. There is (hopefully) no functional change, this is a pure refactor.

These interfaces are quite volatile atm so I didn't bother adding comments as they will be further replaced shortly.

The upcoming `FactSet` implementation is not based on `Fact` (which has terrible locality and probably leads to quite a bit of fragmentation) so this is a step in the right direction.

Pull Request resolved: https://github.com/facebookincubator/Glean/pull/290

Reviewed By: lazamar, malanka

Differential Revision: D40092376

fbshipit-source-id: 86c2632c75c1c47f0c18e3e316f3bb283a834c89