Enforce max symbol limit on document symbols for C++
Summary:
We weren't enforcing symbol limits on client-side generated C++
entities, so the decl to def and external xref facts, expanded on the client
side, could exceed the limit. Without a limit we might see 50k+ symbols in a
file, and we spend a lot of time serializing that, leading too ooms.

Consistently restrict the xref counts on the client using the limit, and enforce the file
limit when not set.

Reviewed By: williamtwilson

Differential Revision: D40133622

fbshipit-source-id: 6c1e675139611b058acbbf3f662fdc9c67dac28a