Property testsuite for the Janitor
Summary:
The test defines a pure `Model` of the janitor that can be updated with a `Command` when something happens:

- A new DB is produced
- A DB download completes
- The sharding assignment changes
- Time elapses
- ... more commands can be added

The test generates a random `[Command]` and applies them in sequence to both the `Model` and the Glean server, comparing the observable state in every step.

The observable state is currently limited to:

- the Catalog
- the restore queue
- (to come next) the active DBs set
- (to come later) the advertised shards
- (to come later) the ShardManager DB counters

TODO: add the testsuite to the OSS Cabal descriptor once this lands.

Reviewed By: simonmar

Differential Revision: D40173156

fbshipit-source-id: b530013e168d53205f2d264d3f85ecf43b2b40aa