Pre-open the whole stack, not just the top
Summary:
We were keeping open the latest DB of each type, but only the topmost
DB in the stack. We also need to keep open the rest of the stack,
otherwise there could be long delays when a client first queries the
DB.

User report:
https://fb.workplace.com/groups/factsaboutcode/posts/1848628748819698/

Verified the suspected root cause by checking the Scuba logs for the
long queries, investigating the logs of the affected server and noting
that the base DB had been opened and closed multiple times.

Reviewed By: pepeiborra

Differential Revision: D40241274

fbshipit-source-id: 520d44d9715ca37f566af8abf46305aa24abbae7