Implement searchRelatedNeighborhood
Summary:
We make 6 distinct types of calls to build api docs reports currently.
This moves that 'report' into Glass, so we can benefit from caching, sharing and reduction in roundtrips from hack -> glass -> glean. There is a 1:1 correspondance from the symbol id to the memcache line after this. And simpler logging as the api docs page load is now 1 call to glass.

- genSearchInheritedFromParents
    - extends
    - parent
    - recursive: true
    - details

- for each parent, by extends, get the containing children
    - Contains
    - Child
    - recursive: false
    - details

- Nth level, containing parents, linear hiearchy
    - Contains
    - Parent
    - recursive: true
    - details

1st order neighborhood:

- 1st level children, contains
   - Contains
   - Child
   - recursive: false
   - details

- 1st level parents, extends
   - Extends
   - Parent
   - recursive: false
   - detailed: true

- 1st level, children, extends
   - Extends
   - Child
   - recursive: false
   - detailed: true

Reviewed By: phlalx

Differential Revision: D40256165

fbshipit-source-id: 35eb8211f37ba58808a7388caeeba585bb99d799