Move Glean.Database.Schema.Transform into Glean.Query.Transform
Summary:
Consolidate a single place for query transformation code.
This is also required because we will need Glean.Query.Codegen to
import things from it, so both files might as well be merged.

All the type checked query transformation code that already existed in `Glean.Query.Transform` will be removed in D40063336.

Reviewed By: simonmar

Differential Revision: D40148674

fbshipit-source-id: 15f5c257dc500a7231248d2ab8e9acdc9b298b7b