Allow values to be transformed at the time of binding
Summary:
When binding a variable we must ensure that the bytes in the variable have the right type. So if a transformation is required we must transform the bytes before putting them in the variable's output.

This is the step we were not taking, and caused the schema transformations issue Michael Park encountered.

This diff adds a constructor to perform this transformation, but doesn't use it anywhere.

This diff makes no functional change.

Reviewed By: simonmar

Differential Revision: D40227317

fbshipit-source-id: 4ac3b2de4475c90e92b8e08394b503680c31acb1