Match ignored record fields during pattern transformation
Summary:
Consider this schema change
```
schema x.1 {
predicate P: { a : nat }
}

schema x.2 {
predicate P: { a : nat, b: maybe nat }
}
```

If the DB has facts of `P.1` we should fill field `b` of `P.2` with `nothing`.
Therefore what should be the result of this query?
```
x.P.2 { a = 1, b = { just = _ } }
```

It mustn't return anything because `b` will always be filled with `nothing`.
But during pattern transformation we throw the `b` pattern away entirely to build a prefix pattern.
To account for that we must not throw the `b` pattern away. We match it against the default value of the field to be ignored during the pattern transformation process.

This means that if a pattern cannot yield results because it will fail to match against the default value of an ignored field, then we won't even go to the database to query for it in the first place.

Reviewed By: simonmar

Differential Revision: D40227316

fbshipit-source-id: bd5809917ffa016cf8f3ee348df27f0dc4ca7464