Enable bytecode transformations
Summary:
This diff turns the switch to enable the new code transformation method.

By removing the transformation code on `flatten` we now get to codegen with untransformed predicates, which triggers the use of the new transformation functions.

The test added exercises the functionality that fails under the previous transformation method.

Reviewed By: simonmar

Differential Revision: D40063335

fbshipit-source-id: 3284d3939cc79e7abfe4ce77a962d1e9c4dcdbf3