More efficient JSON parsing without --use-local-cache
Summary:
Currently we parse the JSON into 1) Folly heap data, marshall into 2) an Aeson `Value`, and then parse into 3) a `JsonFactBatch`. This results in a ton of intermediate allocations and a very high maximum residency which creates problems for large JSON files, as reported here:

https://fb.workplace.com/groups/factsaboutcode/permalink/1845153882500518

I suspect that there is an intermediate state in which two or even all of 1,2,3 are alive in the heap. This Diff does the simplest thing that could possibly work: a pure Haskell JSON parser that fuses steps 2 and 3. The trick is to use `decodeFileStrict'` instead of `decodeFileStrict`.

This reduces memory usage when `--use-local-cache` is not present.

Reviewed By: simonmar

Differential Revision: D40293261

fbshipit-source-id: 63bb9ff28d687ba9284d7a9b8241af342e7ee7d1