Flush DB during finalize
Summary:
Some DBs take a very long time to open (see T134636339). On
investigation this seems to be because these DBs have a lot of
unflushed writes in `.log` files, which have to be replayed when the
DB is opened. Because we open the DB in read-only mode, the writes
never get flushed.

The only way this can happen is if the DB was in read-only mode when
it was backed up (because backup would normally flush first
otherwise), and the only way the DB should be in read-only mode is if
it was closed and re-opened before it got backed up. The only way I
can see that happening is if the backup took a long time to start,
perhaps because there were other DBs already waiting to be backed up,
and the DB got closed due to being idle. I haven't confirmed that this
is the scenario, but if it is then this diff should fix it.

We weren't explicitly flushing the writes, so I'm adding an explicit
flush to the `optimize()` call that we make during finalization of the
DB.

Reviewed By: lazamar

Differential Revision: D40427736

fbshipit-source-id: 6524f1000492f88164027b32a67d60c75c876b5e