Implement clang USR hash to definition lookup
Summary:
We can make clangd a bit smarter by teaching it how to resolve decls that have definitions outside the currently indexed files. ClangD sees an xref and has a USR (symbol hash) of its target. It will be only a target in files already parsed, so will hit a lot of forward decls rather than the remote definitions. We can fix that by asking Glean if it knows.

Will start cautiously here, as ClangD likely can't use lots of fancy things we might provide. To begin: just resolve the go-to-def target location (full range and name range).

This does global decl -> definition resolution via the clang USR hash of a symbol.

- USR -> decl
- decl -> defn, where defn != decl

In the case that the definition site is distinct to the declaration specified by the USR, we return details of the target.
If the hash is invalid or the decl is the same as the definition site, we just throw. To be interpreted by the client as "use what you already have"

Reviewed By: simonmar

Differential Revision: D40461511

fbshipit-source-id: 21e79fc10ddf9f1c213b7335e492730c4db1052a