limit concurrency
Summary:
The Diff Sketcher is firing [~22000 requests](https://fburl.com/scuba/glean_server/dycg5zpr) for a single sketch, all or most of them concurrently. The Prod tier does not have the capacity to serve all these requests before timeout, so the sketcher errors out.

To prevent this, this Diff installs a semaphore limiting the concurrency to 10000 requests. The limit makes little difference in practical performance for the used example. It can be customized by setting the env var `GLEAN_MAX_CONCURRENT_REQUESTS`. A value of 0 will disable it.

Reviewed By: simonmar

Differential Revision: D40464300

fbshipit-source-id: 1fce013838170c3d3ccaf17370b256a114d0c32c