Add static, final, async, transparent, readonly "keyword" attributes
Summary:
New set of extended attributes. These are needed to group symbols in CodeHub API Docs

E.g.
- public static methods
- public static abstract methods
- public abstract methods
- public ..

There's a few of these so use a struct of bool to populate. If not set, false = nothing.

To start with the set is:     Abstract | Final | Async | Static | Readonly | Const | Mutable | Volatile | Virtual | Inline

We can't generate sets dynamically very easily in Angle (or at all really without exhaustively enumerating them), so this just returns a struct of bits.

On the glass side this is turned into a set<enum> so we can simply transfer the valid tokens to the client.

Reviewed By: nhawkes

Differential Revision: D40501907

fbshipit-source-id: db7087a57f09e0c6ae971a101a26b5a66d3a2982