Add static, final, async, abstract,.. attributes to symbol details
Summary:
These will be useful for figuring out if something can be called in an async or static way in Hack, amongst other things.

We return a set of modifier keywords present on the symbol from the enum defined in glass.thrift (and echoed in codemarkup.types.angle)

Presence indicates the modifier keyword is associated with the symbol.

Reviewed By: simonmar

Differential Revision: D40503673

fbshipit-source-id: d81115d07d12b86bd39a216f66dc204e0a0f68a5