model test: improve failure output
Summary:
When the model test fails, it prints *a lot* of `diff`s and a lot of Glean server output. This is because QuickCheck performs shrinking, i.e. runs the test further times with smaller inputs to minimize the example.

Not only the extra `diff`s are confusing, but sometimes the Glean server output is useful to debug a test failure, and this behaviour makes it hard to use it.

This change fixes this by:
1. capturing the std outputs of the Glean server and `diff`
2. only printing the outputs of the final minimized failing example

Reviewed By: simonmar

Differential Revision: D40416563

fbshipit-source-id: 0e6632d318f0148ea1316254d4b91c71c6cdeee6