Support case insensitive scope queries and precise scope lookup
Summary:
Two sets of UX changes:
- case insensitive scope terms (for the first time!), so "vec\map" -> Vec\map accurately
- don't use wildcards for unspecified qname ancestors (e.g. B\C::d won't match *\B\C::d

reasoning for restricting is that people get annoyed that Test\Foo\C\d is hitting searches for Foo\C\d, and arguably in hack, namespaces are so flat and well maintained that if you specify a scope you really mean it.

We might revisit and support '*' in scope terms later.

To implement `cp` we need to distinguish:
- exact match
- case insensitive match
- exact scope
- exact scope with prefix of identifier

Reviewed By: nhawkes

Differential Revision: D40708814

fbshipit-source-id: f4dbaff6fde75a92652f7abf8e933024fd1c0e06