Use the global namespace alias table to quickly resolve canonical aliases
Summary:
We annoyingly get multiple results for scope searches on global aliases as we have a parent wild card for single namespace searches. This is ok in some circumstances, but not for the highly trafficked common names. E.g. "C\count" should not match "Test\Foo\Lib\C\count"

Instead of generating a search for namespace { name = "C", parent = _  } , in these cases we will only look in the resolve global alias.

Reviewed By: nhawkes

Differential Revision: D40707154

fbshipit-source-id: bcd92704ead6d94fae6e43d176c6b8e1dd49ec19