new type predicate.
Summary:
Associate to each type in a signature a "TypeInfo" which contains
1. a pretty printed version of the type (applied auto-import rules)
2. a list of xrefs (spans relative to the pretty-printed representation).

We use a denormalize representation for two reasons
1. efficiency when querying from glass
2. resolution of  symbols within a type are context- dependent so a `type` can't act as a key.

Note: the old `type` field is a textual representation with "full" namespaces.  Rather than re-purposing it for storing the better human representation, I marked it as deprecated and store the new version in typeInfo. Just in case existing clients relied on this particular representation.

```
type Parameter =
  {
    type : maybe Type # deprecated,  old type predicate
    ...
    typeInfo : maybe TypeInfo,  # --> new type predicate
  }
```

Eventually, we need to do this manipulation wherever there's a type field.

Reviewed By: pepeiborra, nhawkes

Differential Revision: D40869718

fbshipit-source-id: 85d3f27280704503e1bd8c7729b851d527a94e17