Rank exact matches and insensitive matches above others
Summary:
Score results and rank them.

This adds a new symbol search result attribute: the `score`.

```
  6: map<string, double> score; // extensible ranking scores
```

And implements a basic "exactness" score, for exact matches and case sensitivitiy.

This is part of implementing "I'm feeling lucky" we want to get a bit better at ranking.

- exact case sensitive results
- above case insensitive results
- above prefix results

We will also need to score these, so other ranking can be combined (e.g. filescore)

This diff adds basic `Ordering` like ranking to put exact matches at the top of each set of results grouped by language/kind.

We can do ranking on the glass side to get the exact matches to the top of the kind box. With `takeFairN` we take a subset of the results, in ranked order. We want to ensure exact matches are at the top of that list.

## Future work

Do smarter scores:
- edit distance
- scope edit distance
- frequency of use
- ...

Reviewed By: nhawkes

Differential Revision: D40788498

fbshipit-source-id: d75856c87a19cec4efd576a672864b696c4d43ff