Add scope search for the empty/global scope
Summary:
We've been missing an important form of scope search: searching global names.

While codehub search can find things  by literal name, for "feeling lucky" `cp`/bunny mode we want to precisely hit the unique name in the global namespace, if the query matches it. The only way to do that (and not pick up other symbols in different scopes but same name) is a global scope query.

We expect to find things like:
- "genmk"
- "C"

which are unambiguous in the Hack global scope.

There's a whole bunch of user requests for search of bare literals to always hit things in the global scope. This is how we'll do this. particularly for `cp genmk` and friends.

Reviewed By: nhawkes

Differential Revision: D40927063

fbshipit-source-id: b04ba8a969354dcaf537eeac127b0ed59e33f097